LPS-158949 Rename file
