+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/TeamMembers/components/TeamMembersTable/hooks/useMyUserAccountByAccountExternalReferenceCode.js
@@ -10,8 +10,8 @@
  */
 
 import {useMemo} from 'react';
-import {useGetMyUserAccount} from '../../../../../../../../common/services/liferay/graphql/user-accounts/queries/useGetMyUserAccount';
-import hasAccountAdministratorRole from '../utils/hasAccountAdministratorRole ';
+import {useGetMyUserAccount} from '../../../../../../../../common/services/liferay/graphql/user-accounts';
+import hasAccountAdministratorRole from '../utils/hasAccountAdministratorRole';
 
 export default function useMyUserAccountByAccountExternalReferenceCode(
 	externalReferenceCode,
